Added global variable so that the modules can access.
This was needed since the original bundle was funciton scoped inside of
webpack, which basically prevented any leakage in the global/window
scope, which was how it was being accessed.  Also separated out the Main
class to its own file and added an App.ts to act as the bootstrap. This
file can be renamed for the specific page or group of pages that share
the same functionality.  Any other "groups" will need their own entry
file.